update my dropdown and cut datalist
I wanted to have the choose a state as an option that popped up almost like a placeholder but i didn't want it to be able to be selected like a real option since it wasn't a state. So I got rid of datalist since disabled and selected attributes aren't compatible with it.